backend/requirements/base.txt
Django==4.2.7
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
markdown==3.5.1
python-docx==0.8.11
PyJWT==2.8.0
cryptography==41.0.5
uvicorn==0.24.0
gunicorn==21.2.0
whitenoise==6.6.0